+++ src/components/card-sort/card-sort.tsx
@@ -1,4 +1,4 @@
-import {CitySortType, CITY_SORTS} from '../../constants/constants.ts';
+import {CitySortType, citySorts} from '../../constants/constants.ts';
 import {useState} from 'react';
 
 interface CardSortProps {
@@ -20,13 +20,13 @@ const CardSort = ({currentSort, onSortChange}: CardSortProps) => {
     <form className="places__sorting" action="#" method="get">
       <span className="places__sorting-caption">Sort by</span>
       <span className="places__sorting-type" tabIndex={0} onClick={toggleIsOpened}>
-        {`${CITY_SORTS.get(currentSort)!.name}`}
+        {`${citySorts.get(currentSort)!.name}`}
         <svg className="places__sorting-arrow" width={7} height={4}>
           <use xlinkHref="#icon-arrow-select"/>
         </svg>
       </span>
       <ul className={`places__options places__options--custom ${(isOpened) ? 'places__options--opened' : ''}`}>
-        {Array.from(CITY_SORTS).map(([sortKey, value]) => (
+        {Array.from(citySorts).map(([sortKey, value]) => (
           <li
             className={`places__option ${(sortKey === currentSort) ? 'places__option--active' : ''}`}
             tabIndex={0}
