+++ src/components/layout/layout.tsx
@@ -1,11 +1,8 @@
 import {Header} from '../header/header.tsx';
 import {Outlet, useLocation} from 'react-router-dom';
 import {Fragment} from 'react';
-import {store} from '../../store/store.ts';
 import {ToastContainer} from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import {fetchFavorites} from '../../store/slices/favorites.ts';
-import {RequestStatus} from '../../constants/constants.ts';
 
 const Layout = () => (
   <Fragment>
@@ -15,12 +12,5 @@ const Layout = () => (
   </Fragment>
 );
 
-const loader = () => {
-  if(store.getState().favorites.requestStatus === RequestStatus.Idle){
-    store.dispatch(fetchFavorites());
-  }
-  return null;
-};
-
-export {Layout, loader};
+export {Layout};
