+++ src/pages/login-page/login-page.tsx
@@ -2,7 +2,7 @@ import {useDocumentTitle} from '../../hooks/use-document-title.ts';
 import {store} from '../../store/store.ts';
 import {login} from '../../store/slices/authentication.ts';
 import {FormEvent, useState} from 'react';
-import {AuthorizationStatus, CITIES} from '../../constants/constants.ts';
+import {AuthorizationStatus, cities} from '../../constants/constants.ts';
 import {Navigate, NavLink} from 'react-router-dom';
 import {LoginCredentials} from '../../contracts/contaracts.ts';
 import {useAppSelector} from '../../hooks/store.ts';
@@ -11,7 +11,7 @@ import {useAppSelector} from '../../hooks/store.ts';
 export const LoginPage = () => {
   const authStatus = useAppSelector((state) => state.authentication.status);
   const [isDisabled, setIsDisabled] = useState(false);
-  const [randomCity] = useState(CITIES[Math.floor(Math.random() * CITIES.length)]);
+  const [randomCity] = useState(cities[Math.floor(Math.random() * cities.length)]);
   useDocumentTitle('Login');
   const handleLogin = (evt: FormEvent) => {
     evt.preventDefault();
